+++ src/c/data.c
@@ -322,6 +322,11 @@ static inline void iot_data_block_free (iot_block_t * block)
 #endif
 }
 
+uint32_t iot_data_hash (const iot_data_t * data)
+{
+  return data ? data->hash : 0u;
+}
+
 static inline void iot_element_free (iot_element_t * element)
 {
   (element->heap) ? free (element) : iot_data_block_free ((iot_block_t*) element);